requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-multipart==0.0.9
pandas==2.3.2
sqlalchemy==2.0.30
aiosqlite==0.19.0
alembic==1.13.1
python-dotenv==1.0.1